+++ src/essarion_build/__init__.py
@@ -82,7 +82,7 @@ from .exceptions import (
     ReasoningFormatError,
 )
 
-__version__ = "0.4.0"
+__version__ = "0.5.0"
 
 __all__ = [
     # Subpackages